Rewrite k, ba to slash commands (#3630)
