+++ src/mahoji/commands/config.ts
@@ -6,7 +6,6 @@ import { ApplicationCommandOptionType, CommandRunOptions } from 'mahoji';
 import { client } from '../..';
 import { BitField, PerkTier, TWEETS_RATELIMITING } from '../../lib/constants';
 import { CombatOptionsArray, CombatOptionsEnum } from '../../lib/minions/data/combatConstants';
-import { GuildSettings } from '../../lib/settings/types/GuildSettings';
 import { removeFromArr, stringMatches } from '../../lib/util';
 import { allAbstractCommands, hasBanMemberPerms, OSBMahojiCommand } from '../lib/util';
 import {
@@ -64,7 +63,7 @@ async function handleTweetsEnable(
 		if (settings.tweetchannel === cID) {
 			return 'Jmod Tweets are already enabled in this channel.';
 		}
-		await mahojiGuildSettingsUpdate(GuildSettings.JModTweets, {
+		await mahojiGuildSettingsUpdate(guild.id, {
 			tweetchannel: cID
 		});
 
@@ -74,7 +73,7 @@ async function handleTweetsEnable(
 		return 'Enabled Jmod Tweets in this channel.';
 	}
 	if (!settings.tweetchannel) return "Jmod Tweets aren't enabled, so you can't disable them.";
-	await mahojiGuildSettingsUpdate(GuildSettings.JModTweets, {
+	await mahojiGuildSettingsUpdate(guild.id, {
 		tweetchannel: null
 	});
 	return 'Disabled Jmod Tweets in this channel.';
@@ -95,7 +94,7 @@ async function handlePetMessagesEnable(
 		if (settings.petchannel) {
 			return 'Pet Messages are already enabled in this guild.';
 		}
-		await mahojiGuildSettingsUpdate(GuildSettings.PetChannel, {
+		await mahojiGuildSettingsUpdate(guild.id, {
 			petchannel: cID
 		});
 		return 'Enabled Pet Messages in this guild.';
@@ -103,7 +102,7 @@ async function handlePetMessagesEnable(
 	if (settings.petchannel === null) {
 		return "Pet Messages aren't enabled, so you can't disable them.";
 	}
-	await mahojiGuildSettingsUpdate(GuildSettings.PetChannel, {
+	await mahojiGuildSettingsUpdate(guild.id, {
 		petchannel: null
 	});
 	return 'Disabled Pet Messages in this guild.';
