+++ src/mahoji/commands/k.ts
@@ -1,0 +1,47 @@
+import { ApplicationCommandOptionType, CommandRunOptions } from 'mahoji';
+
+import { client } from '../..';
+import { PVM_METHODS, PvMMethod } from '../../lib/constants';
+import killableMonsters from '../../lib/minions/data/killableMonsters';
+import { minionKillCommand } from '../lib/abstracted_commands/minionKill';
+import { OSBMahojiCommand } from '../lib/util';
+
+export const killCommand: OSBMahojiCommand = {
+	name: 'k',
+	description: 'Send your minion to kill things.',
+	options: [
+		{
+			type: ApplicationCommandOptionType.String,
+			name: 'name',
+			description: 'The thing you want to kill.',
+			required: true,
+			autocomplete: async value => {
+				return killableMonsters
+					.filter(m => (!value ? true : m.name.toLowerCase().includes(value.toLowerCase())))
+					.map(i => ({ name: i.name, value: i.name }));
+			}
+		},
+		{
+			type: ApplicationCommandOptionType.Number,
+			name: 'quantity',
+			description: 'The amount you want to kill.',
+			required: false,
+			min_value: 0
+		},
+		{
+			type: ApplicationCommandOptionType.String,
+			name: 'method',
+			description: 'If you want to cannon/barrage/burst.',
+			required: false,
+			choices: PVM_METHODS.map(i => ({ name: i, value: i }))
+		}
+	],
+	run: async ({
+		options,
+		userID,
+		channelID
+	}: CommandRunOptions<{ name: string; quantity?: number; method?: PvMMethod }>) => {
+		const user = await client.fetchUser(userID);
+		return minionKillCommand(user, channelID, options.name, options.quantity, options.method ?? 'none');
+	}
+};
